+++ sdk/notificationhubs/notification-hubs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.5 (Unreleased)
+## 1.0.0-beta.6 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,13 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0-beta.5 (2022-09-16)
+
+### Bugs Fixed
+
+- #23222 - `deleteRegistration` not exported on service client.
+- #23218 - single item on registration feed causes error.
 
 ## 1.0.0-beta.4 (2022-09-09)